支持c++20编译。

+++ SOUI/src/core/shostwnd.cpp
@@ -1853,8 +1853,8 @@ void SHostWnd::SHostAnimationHandler::OnNextFrame()
 	STransformation xform;
 	bool bMore = m_pHostWnd->m_hostAnimation->getTransformation(STime::GetCurrentTimeMs(),xform);
 	SMatrix mtx = xform.getMatrix();
-	mtx.preTranslate(-m_rcInit.left,-m_rcInit.top);
-	mtx.postTranslate(m_rcInit.left,m_rcInit.top);
+	mtx.preTranslate((int)-m_rcInit.left, (int)-m_rcInit.top);
+	mtx.postTranslate((int)m_rcInit.left, (int)m_rcInit.top);
 	if(mtx.rectStaysRect())
 	{
 		SRect rc = SRect::IMake(m_rcInit);
